FIX: Refactor mock-time into it's own feature and module
(Thanks Copilot!)
