+++ xlsx/src/import/conditional_formatting.rs
@@ -738,7 +738,7 @@ pub(super) fn load_conditional_formatting(
 }
 
 #[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
+#[allow(clippy::unwrap_used, clippy::expect_used, clippy::panic)]
 mod tests {
     use super::*;
